Moved connection string location and fixed naming
