+++ User Management API/Entities/User.cs
@@ -1,4 +1,4 @@
-﻿namespace User_Management_API.Models;
+﻿namespace User_Management_API.Entities;
 
 public class User
 {
