+++ User Management API/appsettings.Development.json
@@ -1,8 +1,10 @@
 {
   "Logging": {
     "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
+      "Default": "Information"
     }
+  },
+  "ConnectionStrings": {
+    "UserManagementDBConnectionString": "Data Source=(localdb)\\ProjectModels;Initial Catalog=UserManagementDB;Integrated Security=True;Connect Timeout=30;Encrypt=False;Trust Server Certificate=False;Application Intent=ReadWrite;Multi Subnet Failover=False"
   }
 }
