+++ User Management API/appsettings.json
@@ -1,7 +1,4 @@
 {
-  "ConnectionStrings": {
-    "DefaultConnection": "Placeholder;"
-  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
